% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin.R
\name{coin}
\alias{coin}
\title{coin}
\usage{
coin(theta, prior, n, z, alpha = 0.05)
}
\arguments{
\item{theta}{a seq(0,1...) function}

\item{prior}{"uniform" or "triangle"}

\item{n}{number of trials}

\item{z}{number of successes}

\item{alpha}{alpha level}
}
\value{
a named list containing the posterior probabilities, BCI, Posterior mean, and the Prior Mean. Also a plot of the Prior/Post/Likelihood
}
\description{
A function which produces a plot and useful information pertaining to a coin problem in a Bayesian context.
}
\examples{
\dontrun{coin(theta = seq(0,1,length = 101), prior = "uniform", n = 10, z = 4, alpha = 0.05)}
}
