% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycoins.R
\name{shinycoins}
\alias{shinycoins}
\title{shinycoins}
\usage{
shinycoins()
}
\value{
an interactive plot with labels
}
\description{
An interactive app which produces a plot containing the post/prior/likelihood and adjustable widgets for theta, the prior, and the alpha level
}
\examples{
\dontrun{shinycoins()}
}
